Move bans::binary_search_by_name to util
